Adding reset button
Adding a reset button which will relaunch the initialisation routine. I've extracted the actual initialisation code from the starter so we can use that when we reload it. When initialising the KritStorage, the program will now null everything in order to properly reset it on reload. In order to prevent concurrent modification exceptions, we need to prevent the program trying to reset itself mid way through iterating through the krits. Therefore the InspectView will set a flag which, when true, will pause the renderering process and cause a callback to the InspectView which will then reset. When it is done, the renderer will return to normal. 

The InspectView has many more changes as we have moved the layout to a JTabbedPane which allow us to have an Inspect and Manage tab and eventually will give us more control over the layout of the different areas.