Porting to JavaFX and adding graphs
In order to take advantage of Graphs and also to learn a new library, I've ported the windows to JavaFX which I'm really liking. Viewing the profiling from the change I think we may take a minor performance hit but I'm fine with it for the advantages that we get.

And speaking of graphs, included in the new version is four graphs that are updated live as the krits progress showing the amount alive, the number aggressive, the average decay frequency and the average decay rate.