Changing radius calculations to improve aggression speed
After watching the profile logs we were getting up to 17'000'000 calls to the pythagorean distance function in order to calculate the points in a circle effectively. Instead I've switched to a method suggested by someone on StackOverflow which seems to be much quicker and using a massive set of lambda filters we can filter out everything that's not valid which seems a bit quicker. It may not be perfect but it seems to work and does the job well enough.