Cleanup and rethinking the draw methods
Removed the massive block of code that I commented out when testing and have rethunk the drawing methods. Basically now, instead of just rendering the Krit in one function, each Krit will manage a 10x10 image that will be drawn after each update. This means that any form of genetics can also influence the drawing the Krit by modifying the image before the render. This will eventually help the virus implementation as we will be able to make it change the image to have a red border.